+++ .gitignore
@@ -13,10 +13,12 @@ log.txt
 .versions/
 coverage/
 dist/
+dist-test/
+dist-e2e/
 node_modules/
 tmp/
 temp/
-test/
+dist-test/
 $RECYCLE.BIN/
 
 .DS_Store
