+++ package.json
@@ -1,5 +1,5 @@
 {
-  "private": "true",
+  "private": true,
   "name": "ionic2",
   "version": "2.0.0-beta.11",
   "description": "A powerful framework for building mobile and progressive web apps with JavaScript and Angular 2",
@@ -36,18 +36,85 @@
     "es6-shim": "^0.35.0",
     "inquirer": "0.11.0",
     "ionicons": "3.0.0",
+    "lodash": "3.10.1",
     "mkdirp-no-bin": "0.5.1",
+    "q": "1.4.1",
     "reflect-metadata": "^0.1.3",
     "rxjs": "5.0.0-beta.6",
     "zone.js": "^0.6.12"
   },
   "devDependencies": {
+    "@angular/compiler-cli": "0.5.0",
     "@types/jasmine": "^2.2.31",
+    "babel-plugin-transform-es2015-modules-systemjs": "6.9.0",
+    "babel-preset-es2015": "6.9.0",
+    "babel-preset-es2015-native-modules": "6.9.0",
+    "canonical-path": "0.0.2",
+    "connect": "^3.3.4",
+    "conventional-changelog": "^0.5.3",
+    "cpr": "^0.4.1",
+    "del": "~1.1.1",
+    "dgeni": "^0.4.1",
+    "dgeni-packages": "^0.10.18",
+    "event-stream": "^3.3.1",
+    "file-loader": "^0.9.0",
+    "github": "^0.2.4",
+    "glob": "^5.0.14",
     "gulp": "^3.9.1",
+    "gulp-autoprefixer": "^2.3.0",
+    "gulp-babel": "^6.1.2",
     "gulp-cached": "^1.1.0",
+    "gulp-concat": "~2.5.0",
+    "gulp-connect": "^2.2.0",
+    "gulp-conventional-changelog": "^0.7.0",
+    "gulp-if": "^1.2.5",
+    "gulp-insert": "^0.5.0",
+    "gulp-minify-css": "^1.2.2",
+    "gulp-remember": "^0.3.0",
+    "gulp-rename": "~1.2.0",
+    "gulp-sass": "^2.0.4",
+    "gulp-scss-lint": "^0.3.9",
+    "gulp-shell": "^0.4.0",
+    "gulp-strip-debug": "^1.1.0",
+    "gulp-tslint": "^5.0.0",
     "gulp-typescript": "^2.13.6",
-    "lodash": "3.10.1",
-    "typescript": "^2.0.0"
+    "gulp-util": "^3.0.6",
+    "gulp-watch": "^4.2.4",
+    "html-entities": "^1.1.3",
+    "htmlparser2": "^3.8.3",
+    "ionic-cz-conventional-changelog": "^1.0.0",
+    "jasmine-core": "^2.4.1",
+    "jasmine-node": "^1.14.5",
+    "js-yaml": "^3.4.2",
+    "karma": "^0.12.31",
+    "karma-chrome-launcher": "^0.1.7",
+    "karma-jasmine": "^0.3.5",
+    "lazypipe": "^0.2.3",
+    "lunr": "^0.5.12",
+    "merge2": "^0.3.6",
+    "minimist": "^1.1.3",
+    "mkdirp": "^0.5.1",
+    "node-html-encoder": "0.0.2",
+    "node-libs-browser": "^0.5.2",
+    "request": "2.53.0",
+    "rollup": "^0.34.3",
+    "rollup-plugin-commonjs": "^3.3.1",
+    "rollup-plugin-node-resolve": "^2.0.0",
+    "run-sequence": "^1.1.0",
+    "sassdoc": "^2.1.20",
+    "semver": "^5.0.1",
+    "serve-static": "^1.9.2",
+    "source-map-support": "^0.2.10",
+    "strip-sourcemap-loader": "0.0.1",
+    "systemjs": "0.19.6",
+    "through2": "^0.6.3",
+    "tslint": "^3.10.2",
+    "tslint-ionic-rules": "0.0.2",
+    "typescript": "^2.0.0",
+    "vinyl": "^0.4.6",
+    "webpack": "^2.1.0-beta.20",
+    "webpack-dev-server": "^1.14.1",
+    "yargs": "^3.6.0"
   },
   "config": {
     "commitizen": {
