+++ scripts/e2e/e2e.template.js
@@ -1,7 +1,7 @@
 describe('<%= relativePath %>: <%= platform %>', function() {
 
 it('should init', function() {
-  browser.get('http://localhost:<%= buildConfig.protractorPort %>/dist/e2e/<%= relativePath %>/index.html?ionicplatform=<%= platform %>&ionicOverlayCreatedDiff=0&ionicanimate=false&snapshot=true');
+  browser.get('http://localhost:<%= buildConfig.protractorPort %>/dist-e2e/e2e/<%= relativePath %>/index.html?ionicplatform=<%= platform %>&ionicOverlayCreatedDiff=0&ionicanimate=false&snapshot=true');
 });
 
 <%= contents %>
