+++ scripts/e2e/webpackEntryPoints.json
@@ -1,122 +1,1 @@
-{
-  "test/action-sheet/basic/index": "./test/action-sheet/basic/entry",
-  "test/alert/basic/index": "./test/alert/basic/entry",
-  "test/alert/dismiss/index": "./test/alert/dismiss/entry",
-  "test/app/animations/index": "./test/app/animations/entry",
-  "test/app/cordova/index": "./test/app/cordova/entry",
-  "test/app/gesture-collision/index": "./test/app/gesture-collision/entry",
-  "test/app/gestures/index": "./test/app/gestures/entry",
-  "test/app/storage/index": "./test/app/storage/entry",
-  "test/app/typography/index": "./test/app/typography/entry",
-  "test/badge/basic/index": "./test/badge/basic/entry",
-  "test/button/basic/index": "./test/button/basic/entry",
-  "test/button/block/index": "./test/button/block/entry",
-  "test/button/clear/index": "./test/button/clear/entry",
-  "test/button/dynamic/index": "./test/button/dynamic/entry",
-  "test/button/fab/index": "./test/button/fab/entry",
-  "test/button/full/index": "./test/button/full/entry",
-  "test/button/icons/index": "./test/button/icons/entry",
-  "test/button/outline/index": "./test/button/outline/entry",
-  "test/button/raised/index": "./test/button/raised/entry",
-  "test/button/round/index": "./test/button/round/entry",
-  "test/button/sizes/index": "./test/button/sizes/entry",
-  "test/card/advanced/index": "./test/card/advanced/entry",
-  "test/card/basic/index": "./test/card/basic/entry",
-  "test/card/images/index": "./test/card/images/entry",
-  "test/card/list/index": "./test/card/list/entry",
-  "test/card/map/index": "./test/card/map/entry",
-  "test/card/social/index": "./test/card/social/entry",
-  "test/checkbox/basic/index": "./test/checkbox/basic/entry",
-  "test/chip/basic/index": "./test/chip/basic/entry",
-  "test/chip/delete/index": "./test/chip/delete/entry",
-  "test/chip/icon/index": "./test/chip/icon/entry",
-  "test/chip/image/index": "./test/chip/image/entry",
-  "test/content/basic/index": "./test/content/basic/entry",
-  "test/content/fullscreen/index": "./test/content/fullscreen/entry",
-  "test/datetime/basic/index": "./test/datetime/basic/entry",
-  "test/datetime/labels/index": "./test/datetime/labels/entry",
-  "test/grid/alignment/index": "./test/grid/alignment/entry",
-  "test/grid/basic/index": "./test/grid/basic/entry",
-  "test/grid/full/index": "./test/grid/full/entry",
-  "test/icon/basic/index": "./test/icon/basic/entry",
-  "test/infinite-scroll/basic/index": "./test/infinite-scroll/basic/entry",
-  "test/infinite-scroll/short-list/index": "./test/infinite-scroll/short-list/entry",
-  "test/input/clear-input/index": "./test/input/clear-input/entry",
-  "test/input/fixed-inline-labels/index": "./test/input/fixed-inline-labels/entry",
-  "test/input/floating-labels/index": "./test/input/floating-labels/entry",
-  "test/input/form-inputs/index": "./test/input/form-inputs/entry",
-  "test/input/highlight/index": "./test/input/highlight/entry",
-  "test/input/inline-labels/index": "./test/input/inline-labels/entry",
-  "test/input/input-focus/index": "./test/input/input-focus/entry",
-  "test/input/inset-inputs/index": "./test/input/inset-inputs/entry",
-  "test/input/placeholder-labels/index": "./test/input/placeholder-labels/entry",
-  "test/input/stacked-labels/index": "./test/input/stacked-labels/entry",
-  "test/item/buttons/index": "./test/item/buttons/entry",
-  "test/item/dividers/index": "./test/item/dividers/entry",
-  "test/item/groups/index": "./test/item/groups/entry",
-  "test/item/icons/index": "./test/item/icons/entry",
-  "test/item/images/index": "./test/item/images/entry",
-  "test/item/media/index": "./test/item/media/entry",
-  "test/item/reorder/index": "./test/item/reorder/entry",
-  "test/item/sliding/index": "./test/item/sliding/entry",
-  "test/item/text/index": "./test/item/text/entry",
-  "test/list/header-scenarios/index": "./test/list/header-scenarios/entry",
-  "test/list/headers/index": "./test/list/headers/entry",
-  "test/list/inset/index": "./test/list/inset/entry",
-  "test/list/no-lines/index": "./test/list/no-lines/entry",
-  "test/list/repeat-headers/index": "./test/list/repeat-headers/entry",
-  "test/list/sticky/index": "./test/list/sticky/entry",
-  "test/loading/basic/index": "./test/loading/basic/entry",
-  "test/loading/tabs/index": "./test/loading/tabs/entry",
-  "test/menu/basic/index": "./test/menu/basic/entry",
-  "test/menu/disable-swipe/index": "./test/menu/disable-swipe/entry",
-  "test/menu/enable-disable/index": "./test/menu/enable-disable/entry",
-  "test/menu/multiple/index": "./test/menu/multiple/entry",
-  "test/menu/overlay/index": "./test/menu/overlay/entry",
-  "test/menu/push/index": "./test/menu/push/entry",
-  "test/menu/reveal/index": "./test/menu/reveal/entry",
-  "test/modal/basic/index": "./test/modal/basic/entry",
-  "test/nav/basic/index": "./test/nav/basic/entry",
-  "test/nav/child-navs/index": "./test/nav/child-navs/entry",
-  "test/nav/init-async/index": "./test/nav/init-async/entry",
-  "test/nav/insert-views/index": "./test/nav/insert-views/entry",
-  "test/nav/memory/index": "./test/nav/memory/entry",
-  "test/nav/nested/index": "./test/nav/nested/entry",
-  "test/picker/basic/index": "./test/picker/basic/entry",
-  "test/popover/basic/index": "./test/popover/basic/entry",
-  "test/radio/basic/index": "./test/radio/basic/entry",
-  "test/range/basic/index": "./test/range/basic/entry",
-  "test/refresher/basic/index": "./test/refresher/basic/entry",
-  "test/scroll/basic/index": "./test/scroll/basic/entry",
-  "test/searchbar/basic/index": "./test/searchbar/basic/entry",
-  "test/searchbar/nav/index": "./test/searchbar/nav/entry",
-  "test/searchbar/toolbar/index": "./test/searchbar/toolbar/entry",
-  "test/segment/basic/index": "./test/segment/basic/entry",
-  "test/segment/nav/index": "./test/segment/nav/entry",
-  "test/segment/swipe/index": "./test/segment/swipe/entry",
-  "test/select/multiple-value/index": "./test/select/multiple-value/entry",
-  "test/select/single-value/index": "./test/select/single-value/entry",
-  "test/show-hide-when/basic/index": "./test/show-hide-when/basic/entry",
-  "test/slides/basic/index": "./test/slides/basic/entry",
-  "test/slides/controller/index": "./test/slides/controller/entry",
-  "test/slides/intro/index": "./test/slides/intro/entry",
-  "test/slides/loop/index": "./test/slides/loop/entry",
-  "test/slides/scroll/index": "./test/slides/scroll/entry",
-  "test/spinner/basic/index": "./test/spinner/basic/entry",
-  "test/tabs/advanced/index": "./test/tabs/advanced/entry",
-  "test/tabs/badges/index": "./test/tabs/badges/entry",
-  "test/tabs/basic/index": "./test/tabs/basic/entry",
-  "test/tabs/child-navs/index": "./test/tabs/child-navs/entry",
-  "test/tabs/colors/index": "./test/tabs/colors/entry",
-  "test/tabs/ghost/index": "./test/tabs/ghost/entry",
-  "test/tabs/tab-bar-scenarios/index": "./test/tabs/tab-bar-scenarios/entry",
-  "test/toast/basic/index": "./test/toast/basic/entry",
-  "test/toggle/basic/index": "./test/toggle/basic/entry",
-  "test/toolbar/colors/index": "./test/toolbar/colors/entry",
-  "test/toolbar/scenarios/index": "./test/toolbar/scenarios/entry",
-  "test/virtual-scroll/basic/index": "./test/virtual-scroll/basic/entry",
-  "test/virtual-scroll/cards/index": "./test/virtual-scroll/cards/entry",
-  "test/virtual-scroll/image-gallery/index": "./test/virtual-scroll/image-gallery/entry",
-  "test/virtual-scroll/sliding-item/index": "./test/virtual-scroll/sliding-item/entry",
-  "test/virtual-scroll/variable-size/index": "./test/virtual-scroll/variable-size/entry"
-}
+{}
