+++ src/components/alert/test/dismiss/AppModule.ts
@@ -1,5 +1,5 @@
 import { Component, NgModule } from '@angular/core';
-import { IonicApp, IonicModule, AlertController, LoadingController, NavController } from '../../../dist';
+import { IonicApp, IonicModule, AlertController, LoadingController, NavController } from '../../../dist-e2e';
 import { FormBuilder, ControlGroup, Validators } from '@angular/common';
 import { FormGroup } from '@angular/forms';
